Move all selector classes to the top level package
