+++ audioswitch/src/main/java/com/twilio/audioswitch/AudioDeviceChangeListener.kt
@@ -1,4 +1,4 @@
-package com.twilio.audioswitch.selection
+package com.twilio.audioswitch
 
 /** A listener that receives the most recently available [AudioDevice]s.
  * Receives a list of the most recently available [AudioDevice]s. Also provides the
